src/components/Form.js: moved form to Form dir, updated imports
